Workaround https://github.com/OpenTTD/OpenTTD/issues/10549
If a GRF without parameters is upgraded to one with parameters, then the
default values are (sometimes?) ignored and the parameters instead take
the value 0.

This commit makes the default values for all parameters 0 so that people
affected by this bug still get the intended defaults.

An unfortunate side-effect is that people already on version 0.7 will
silently have their parameters changed (or perhaps not silently if their
values were set above the new max_value for the parameter).